tgbot/requirements.txt
aiogram>=3.0.0
aiohttp
httpx
pydantic
python-dotenv
